[fix] not able to delete root accidentally

+++ src/Components/ViewModels/PasswordViewModel.ts
@@ -137,6 +137,9 @@ export const usePasswortViewModel = (automergeFacade: AutomergeFacade) => {
     }
 
     function deleteItem(item: Item) {
+        if (item.id === "") {
+            return;
+        }
         reactiveFacade.deleteItem(item.id);
         item.deleted = true;
         setCurItem(getRootFolder());
